requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorlog==6.7.0
frozenlist==1.3.3
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
motion-python==0.1.62
multidict==6.0.4
numpy==1.24.3
openai==0.27.8
pandas==2.0.2
psutil==5.9.5
pydantic==1.10.9
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
rich==13.4.1
six==1.16.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
